fixed the payments methods amount to reflect correctly the ui price

+++ app/src/main/java/com/oliviermarteaux/a049_joiefull/ui/navigation/NavGraph.kt
@@ -27,7 +27,7 @@ fun JoiefullNavHost(
     modifier: Modifier = Modifier,
     innerPadding: PaddingValues,
     payUiState: PaymentUiState,
-    onGooglePayButtonClick: () -> Unit,
+    onGooglePayButtonClick: (Double) -> Unit,
 ) {
     NavHost(
         navController = navController,
