+++ app/src/main/java/com/oliviermarteaux/a049_joiefull/ui/screens/home/HomeScreen.kt
@@ -88,7 +88,7 @@ fun HomeScreen(
     viewModel: HomeViewModel,
     contentType: SharedContentType = SharedContentType.LIST_ONLY,
     payUiState: PaymentUiState = PaymentUiState.NotStarted,
-    onGooglePayButtonClick: () -> Unit,
+    onGooglePayButtonClick: (Double) -> Unit,
 ) {
     val uiState = viewModel.uiState
